fix(homepage) minor improvements for Helm users (#18729)
**Description**
Small fixes and clean up things to make the Homepage chart easier to
install for Helm users. This seems to work fine for me installing with
Helm, and I think it should work for all platforms.

I think this is a breaking change for Helm users and I added a bit to
the Changelog to explain it too. Let me know and I can format that
differently. For right now it's just a link to the repository since the
PR doesn't exist yet.

The bugfix is fixing the path on force-services-config. The rest is a
bit of templating to make things easier to swap between editing the
config in the configmaps vs in the container.

⚒️ Fixes  # N/A

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [X] 🪛 Bugfix
- [X] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [X] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
I ran this values.yaml file in my own k3s cluster using Helm with
ingress added. I didn't repackage the chart to test locally or anything,
but I think testing the values.yaml file is enough. Let me know and I
can do some more testing.

**📃 Notes:**
I've been learning k8s and Helm recently for fun and found some of
these. Any feedback is appreciated.

**✔️ Checklist:**

- [ ] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [X] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [X] ⬆️ I increased versions for any altered app according to semantic
versioning

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: Kjeld Schouten <info@kjeldschouten.nl>
Co-authored-by: Kjeld Schouten <kjeld@schouten-lebbing.nl>