fix(whisper-asr-webservice): liveness probe fix, add large-v3 model, update version, remove unused storage (#20042)
**Description**

- Update of the container images, as they were outdated (v1.2.3 ->
v1.3.0)
- Change of the `liveness probe` from `http`-> `tcp`. Because now the
pod is killed during transcoding/translating operations due to the
unresponsiveness of the http interface during this time.
- add whisper `large-v3` model as option
- removing `/root/.cache/faster_whisper` storage values/questions as
there is no use for it as all models are stored in
`/root/.cache/whisper`
- Some minor description fixes
⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [X] ⚙️ Feature/App addition
- [X] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

**📃 Notes:**
For reference about the "pod gets killed issue" with
whisper-asr-webservice if `liveness probe` is set to `http`
https://discord.com/channels/830763548678291466/1184211864700338186

**✔️ Checklist:**

- [X] ⚖️ My code follows the style guidelines of this project
- [X] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [X] ⬆️ I increased versions for any altered app according to semantic
versioning
- [X] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):` or `chore(chart-name):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Signed-off-by: v3DJG6GL <72495210+v3DJG6GL@users.noreply.github.com>