feat!: Add integrated Volsync support and remove Velero Support (#790)
**Description**
Velero has a whole host of issues:
1. It does not work with RWO blockdevice PVCs
2. It does not play nice with GitOps (as you don't want all manifests
ported)
3. It does not play nice with TrueNAS SCALE (which iX is to blame for)

But combined: It sucks and only works if you fit within a VERY tight
design-specification for your clustered resources.
So while we spend many ours on it: It has to go.

Introducing: VolSync support.

It just does storage backups and does it well.
If you reinstall with the same VolSync setup, it fetches your data from
S3 storage before starting stuff.
Just works (tm).

Also should just work(tm) with TrueNAS SCALE.

**⚙️ Type of change**

- [x] ⚙️ Feature/App addition
- [ ] 🪛 Bugfix
- [x] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
<!--
Please describe the tests that you ran to verify your changes. Provide
instructions so we can reproduce. Please also list any relevant details
for your test configuration
-->

**📃 Notes:**
<!-- Please enter any other relevant information here -->

**✔️ Checklist:**

- [ ] ⚖️ My code follows the style guidelines of this project
- [ ] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [ ] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [ ] ⬆️ I increased versions for any altered app according to semantic
versioning
- [ ] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):` or `chore(chart-name):`

**➕ App addition**

If this PR is an app addition please make sure you have done the
following.

- [ ] 🖼️ I have added an icon in the Chart's root directory called
`icon.png`

---

_Please don't blindly check all the boxes. Read them and only check
those that apply.
Those checkboxes are there for the reviewer to see what is this all
about and
the status of this PR with a quick glance._

---------

Co-authored-by: Stavros kois <s.kois@outlook.com>