fix(qbittorrent): allow webui access from all IP families (#21664)
**Description**
<!--
Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.
-->

Currently, the Qbittorrent's WebGUI Address is hardcoded to "0.0.0.0"
which only allows access to the WebGUI via IPv4 addresses.
When the App is created with the Host Network option, IPv6 addresses
might becorme available and users might want to be able to access the
WebUI via those addresses.
Taking that scenario into account, change the hardcoded default created
by commit 68b93f4 to qbittorrent's default "*", allowing all IP address
families to access the WebGUI.

⚒️ Fixes  # <!--(issue)-->

**⚙️ Type of change**

- [ ] ⚙️ Feature/App addition
- [x] 🪛 Bugfix
- [ ] ⚠️ Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] 🔃 Refactor of current code

**🧪 How Has This Been Tested?**
Tested the change in a running app to see if the WebUI was still
available.

**📃 Notes:**
N/A

**✔️ Checklist:**

- [ ] ⚖️ My code follows the style guidelines of this project
- [x] 👀 I have performed a self-review of my own code
- [ ] #️⃣ I have commented my code, particularly in hard-to-understand
areas
- [ ] 📄 I have made corresponding changes to the documentation
- [x] ⚠️ My changes generate no new warnings
- [ ] 🧪 I have added tests to this description that prove my fix is
effective or that my feature works
- [x] ⬆️ I increased versions for any altered app according to semantic
versioning
- [x] I made sure the title starts with `feat(chart-name):`,
`fix(chart-name):` or `chore(chart-name):`

Signed-off-by: Tiago Gaspar <tiagogaspar8@gmail.com>